Update buildTools.ps1.
Stop always expanding packages.

Always expanding the packages made it easier to make sure that the cache
was not corrupted just by checking the sha512 of the nupkg and then
expanding the nupkg.

However, this is causing conflicts with ghostProtocol.ps1 use of
Add-Type.

This change should fix builds that use
    powershell.exe
but builds that use
    pwsh.exe or pwsh
will need more work.

****************************************************************************
** Revert this change if Add-Type is no longer used in ghostProtocol.ps1. **
****************************************************************************